make chatgpt response tests more robust
Even when given specific instructions like "Return the response 'Hi'."
ChatGPT (and LLMs in general) don't always follow the
instructions *exactly* (for example, ChatGPT will often respond to the above
prompt with "Hi!").

As a result, equality testing on the response makes for flaky tests. This change
keeps the test prompts, but instead matches on the responses with `=~`. Still
not perfect, but less likely to be flaky, which in tests seems like a win.